+++ packages/dart/vault_data_manager/test/fixtures/iam_api_service_fixtures.dart
@@ -1,4 +1,5 @@
-import 'package:affinidi_tdk_iam_client/affinidi_tdk_iam_client.dart';
+import 'package:affinidi_tdk_consumer_iam_client/affinidi_tdk_consumer_iam_client.dart'
+    as consumer_iam;
 import 'package:affinidi_tdk_vault/affinidi_tdk_vault.dart';
 import 'package:dio/dio.dart';
 
@@ -6,12 +7,12 @@ class IamApiServiceFixtures {
   static const String testGranteeDid = 'did:test:123';
   static const Permissions testPermissions = Permissions.read;
 
-  static Response<GrantAccessOutput> get successfulGrantAccessResponse =>
-      Response(
-        data: GrantAccessOutput((b) => b..success = true),
-        statusCode: 200,
-        requestOptions: RequestOptions(),
-      );
+  static Response<consumer_iam.UpdateAccessOutput>
+      get successfulUpdateAccessResponse => Response(
+            data: consumer_iam.UpdateAccessOutput((b) => b..success = true),
+            statusCode: 200,
+            requestOptions: RequestOptions(),
+          );
 
   static Response<dynamic> get successfulRevokeAccessResponse => Response(
         data: null,