+++ libs/dart/vault_flutter_utils/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.11.3
+
+### Changes
+
+---
+
+Chore: Dependencies Update 
+
+---
+
 ## 1.11.2
 
 ### Changes